push last

+++ user-service-domain/src/main/java/com/test/domain/user/spi/IUserRepository.java
@@ -1,19 +1,17 @@
 package com.test.domain.user.spi;
 
-import com.test.domain.user.api.model.IUser;
-
 import java.util.List;
 import java.util.Optional;
 
 public interface IUserRepository<TU extends IUser> {
 
-    Optional<TU> find(int id);
+    Optional<TU> find(long id);
 
     List<TU> findAll();
 
-    int create(TU user);
+    long create(TU user);
 
-    void update(int id, TU user);
+    void update(long id, TU user);
 
-    void delete(int id);
+    void delete(long id);
 }
